Refactor add_book into module
