+++ data/book.json
@@ -1,1 +1,1 @@
-[{"id":"bca0aab990eb6e6e9981","publish_date":null,"archived":false,"source":null,"publisher":"Macaw co","cover_state":"good"}]
+[{"id":"bd7cd6e0fb60ee55b8b7","publish_date":null,"archived":false,"source":null,"publisher":"Macaw co","cover_state":"good"}]
