+++ data/label.json
@@ -3,5 +3,10 @@
     "id": "e29afca0-aa59-013b-a8b5-07a5d4b0e88a",
     "title": "The Great Gatsby",
     "color": "red"
+  },
+  {
+    "id": "77691132a55a4c9b7e34",
+    "title": "fasd",
+    "color": "blue"
   }
 ]
