+++ lib_for_main/app.rb
@@ -4,9 +4,11 @@ require_relative '../lib/books/label'
 require_relative '../lib/handle_data'
 require_relative '../lib/game'
 require 'json'
-# require 'uuid'
+require_relative '../lib/modules/add_book'
 
 class App
+  include AddBook
+
   def initialize
     books_data = HandleData.new('book')
     @books = books_data.read.map do |book|
@@ -59,25 +61,6 @@ class App
 
   def source_list
     puts @sources
-  end
-
-  def add_book
-    puts "Let's create a book"
-    puts "Can we have this book's title?"
-    title = gets.chomp
-    puts "Can we have this book's publish date?"
-    p_date = gets.chomp
-    puts "Can we have this book's publisher?"
-    publisher = gets.chomp
-    puts "Can we have this book's cover state?"
-    c_state = gets.chomp
-    puts "Can we have this book's cover color?"
-    c_color = gets.chomp
-    id = UUID.generate
-    label = Label.new(id: id, title: title, color: c_color)
-    book = Book.new(id: id, publish_date: p_date, publisher: publisher, cover_state: c_state, archived: false)
-    save_book(book, label)
-    puts 'Book added successfully'
   end
 
   def add_album
